implement squareMode based on cell width/height, workaround for saveImage in squareMode
New logic for square mode size calculations that are based on
individual cell width and height so now square mode will always have
square cells even if number of x and y items are different. Also fixed
extra whitespace in saveImage by over-riding saveImage and temporarily
adjusting the width and height of the chart in order to remove extra
space.